+++ h8/drawcircle.h
@@ -13,7 +13,7 @@ void Circle::draw( )
     c.set( );
     glBegin( GL_POLYGON );
     for ( int i = 0; i < CIRCLE_LENGTH; i++ )
-    { glVertex2f( p1.getx( ) + r * cos( 2 * PI / CIRCLE_LENGTH * i ), p1.gety( ) + r * sin( 2 * PI / CIRCLE_LENGTH * i ) ); }
+    { glVertex2f( p1.getx( ) + (float)r * (float)cos( 2 * PI / CIRCLE_LENGTH * i ), p1.gety( ) + (float)r * (float)sin( 2 * PI / CIRCLE_LENGTH * i ) ); }
     glEnd( );
     return;
 }
